+++ apps/app_test.go
@@ -1,3 +1,5 @@
+// +build !e2e
+
 package apps
 
 import (
