+++ server/api/impl/appservices/kv_test.go
@@ -1,3 +1,5 @@
+// +build !e2e
+
 package appservices
 
 import (
