+++ server/api/impl/proxy/bindings_test.go
@@ -1,3 +1,5 @@
+// +build !e2e
+
 package proxy
 
 import (
