+++ server/api/impl/store/subscriptions_test.go
@@ -1,3 +1,5 @@
+// +build !e2e
+
 package store
 
 import (
