+++ server/http/restapi/main_test.go
@@ -1,6 +1,8 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+// +build e2e
+
 package restapi
 
 import (
