+++ server/utils/httputils/utils_test.go
@@ -1,5 +1,6 @@
 // Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
+// +build !e2e
 
 package httputils
